Move TestValidator into its own module
