+++ cli/tests/stake.rs
@@ -10,7 +10,7 @@ use solana_cli::{
     test_utils::{check_ready, check_recent_balance},
 };
 use solana_client::rpc_client::RpcClient;
-use solana_core::validator::{TestValidator, TestValidatorOptions};
+use solana_core::test_validator::{TestValidator, TestValidatorOptions};
 use solana_faucet::faucet::run_local_faucet;
 use solana_sdk::{
     account_utils::StateMut,
