+++ cli/tests/vote.rs
@@ -5,7 +5,7 @@ use solana_cli::{
     test_utils::check_recent_balance,
 };
 use solana_client::rpc_client::RpcClient;
-use solana_core::validator::TestValidator;
+use solana_core::test_validator::TestValidator;
 use solana_faucet::faucet::run_local_faucet;
 use solana_sdk::{
     account_utils::StateMut,
