+++ tokens/tests/commands.rs
@@ -1,5 +1,5 @@
 use solana_banks_client::start_tcp_client;
-use solana_core::validator::{TestValidator, TestValidatorOptions};
+use solana_core::test_validator::{TestValidator, TestValidatorOptions};
 use solana_sdk::native_token::sol_to_lamports;
 use solana_tokens::commands::test_process_distribute_tokens_with_client;
 use std::fs::remove_dir_all;
